Utils: Fix MSVC warning in MultitextCursor
warning C4267: 'return': conversion from 'size_t' to 'int', possible loss of data

+++ src/libs/utils/multitextcursor.cpp
@@ -90,7 +90,7 @@ bool MultiTextCursor::hasMultipleCursors() const
 
 int MultiTextCursor::cursorCount() const
 {
-    return m_cursors.size();
+    return static_cast<int>(m_cursors.size());
 }
 
 void MultiTextCursor::movePosition(QTextCursor::MoveOperation operation,
